requirements-dev.txt
coverage
pytest
pytest-cov
pytest-mock
mypy
pylint
black
coverage
python-dotenv